// src/app/interfaces/PatientInfo.ts
export interface PatientInfo {
  nom: string;
  prenom: string;
  email: string;
  tel?: string;
  phone?: string;
  age?: number;
  height?: number;
  weight?: number;
  goal?: string;
  lifestyleLevel?: string;
  ville?: string;
  dateNaissance?: string;
  sexe?: string;
  adresse?: string;
  typeAbonnement?: string;
}